+++ react/modules/seller.ts
@@ -1,15 +1,15 @@
-import type { ProductSummaryTypes } from 'vtex.product-summary-context'
+import type { ProductTypes } from 'vtex.product-context'
 
-export function getFirstAvailableSeller(
-  sellers?: ProductSummaryTypes.Seller[]
-) {
+export function getDefaultSeller(sellers?: ProductTypes.Seller[]) {
   if (!sellers || sellers.length === 0) {
     return
   }
 
-  const availableSeller = sellers.find(
-    (seller) => seller.commertialOffer.AvailableQuantity !== 0
-  )
+  const defaultSeller = sellers.find((seller) => seller.sellerDefault)
 
-  return availableSeller
+  if (!defaultSeller) {
+    return sellers[0]
+  }
+
+  return defaultSeller
 }